rename `slack-share-icon` to `slackShareIcon`
